REF: relocate comparison utilities to util.py
